added navbar to sign in/signup/portfolio home and made navbar be logged in/loggedout unique
