+++ client/src/pages/Portfolio/Home/PortfolioHome.jsx
@@ -1,6 +1,7 @@
-import React, { useEffect } from "react";
+import React, { useEffect, Fragment } from "react";
 import User from "../../../utils/Stores/User"
 import API from "../../../utils/Stores/Cryptos/cryptoAPI"
+import AppBar from "../../../components/AppBar/AppBar"
 
 export default function(){
     User.refreshOnLoad();
@@ -16,8 +17,11 @@ export default function(){
     }, [])
     
     return(
-        <div>
-            this will be the member page
-        </div>
+        <Fragment>
+            <AppBar/>
+            <div>
+                this will be the member page
+            </div>
+        </Fragment>
     )
 }
